Unifying LPC motd/init

+++ cmslpc-notebook/config/bootstrap.sh
@@ -1,3 +1,12 @@
+echo ===== Creating base directories for NFS mounts
+mkdir -p /uscmst1b_scratch/lpc1
+mkdir -p /uscms_data/d1
+mkdir -p /uscms_data/d2
+mkdir -p /uscms_data/d3
+mkdir -p /uscms
+
+
+echo ===== Putting HTCondor wrappers in place
 declare -a REPLACE_CONDOR_CMD=('condor_q' 'condor_rm' 'condor_qedit' 'condor_tail' 'condor_submit' 'condor_history')
 declare -a REPLACE_CONDOR_RM_LINKS=('condor_vacate_job' 'condor_suspend' 'condor_continue')
 declare -a DELETE_CONDOR_RM_LINKS=('condor_hold' 'condor_release')
@@ -10,24 +19,36 @@ cp /tmp/gco_scripts/htcondor/cmslpc-local-conf.py /usr/local/bin/cmslpc-local-co
 # Replacing various HTCondor commands with CSI/GCO wrappers
 for i in "${REPLACE_CONDOR_CMD[@]}"
 do
-  mv /usr/bin/${i} /usr/bin/original_${i}
+  if [ -e /usr/bin/${i} ]; then 
+    mv /usr/bin/${i} /usr/bin/original_${i}
+  else
+    :
+  fi
 done 
 
 for j in "${REPLACE_CONDOR_RM_LINKS[@]}"
 do 
-  ln -sf /usr/bin/original_condor_rm /usr/bin/${j}
+  if [ -e /usr/bin/${i} ]; then 
+    :
+  else
+    ln -sf /usr/bin/original_condor_rm /usr/bin/${j}
+  fi
 done
 
 for k in "${DELETE_CONDOR_RM_LINKS[@]}"
 do
-  rm -f /usr/bin/${k} && cp /usr/bin/original_condor_rm /usr/bin/original_${k}
+ if [ -h /usr/bin/${i} ]; then 
+    rm -f /usr/bin/${k} && cp /usr/bin/original_condor_rm /usr/bin/original_${k}
+  else
+    :
+  fi
 done
 
 for l in "${NAME_FLAG_CHK_CMD[@]}"
 do
   cp /tmp/gco_scripts/htcondor/cmslpc-name-flag-check.py /usr/local/bin/${l} && chmod +755 /usr/local/bin/${l}
   # Creating symlinks for these binaries under /usr/libexec/condor. Needed to preserve original `condor_rm` functionality
-  cp /usr/bin/original_${l} /usr/libexec/condor/${l} && chmod +755 /usr/local/bin/${l}
+  ln -sf /usr/bin/original_${l} /usr/libexec/condor/${l}
 done
 
 cp /tmp/gco_scripts/htcondor/cmslpc-condor-q.py /usr/local/bin/condor_q && chmod +755 /usr/local/bin/condor_q
@@ -35,8 +56,11 @@ cp /tmp/gco_scripts/htcondor/cmslpc-condor-submit.py /usr/local/bin/condor_submi
 
 rm -rf /tmp/gco_scripts
 
+mkdir -p /etc/condor/certs
 git clone https://hepcloud-git.fnal.gov/ECF-GCO-public/htcondor-config-files.git /tmp/htcondor-config-files
 cp /tmp/htcondor-config-files/service_configs/*cmslpc_interactive* /etc/condor/config.d/
 cp /tmp/htcondor-config-files/mapfiles/cmslpc.condor_mapfile /etc/condor/certs/condor_mapfile
 
 rm -rf /tmp/htcondor-config-files
+
+$(condor_config_val MASTER)
